requirements.txt
streamlit
langchain
langchain-community
sentence-transformers
chromadb
torch
transformers
python-dotenv
